streamlit_app/requirements.txt
streamlit==1.30.0
elasticsearch==8.10.0
pandas==2.1.4
pillow==10.2.0
python-dotenv==1.0.0
